requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
brotlipy==0.7.0
certifi==2021.10.8
cffi @ file:///C:/ci_310/cffi_1642682485096/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.3
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conda==4.12.0
conda-content-trust @ file:///tmp/build/80754af9/conda-content-trust_1617045594566/work
conda-package-handling @ file:///C:/ci/conda-package-handling_1649106011304/work
cryptography @ file:///C:/ci/cryptography_1639414720302/work
distlib==0.3.6
Django==4.1.6
django-crispy-forms==2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
filelock==3.9.0
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
menuinst @ file:///C:/ci/menuinst_1631733438520/work
Pillow==9.4.0
platformdirs==3.0.0
psycopg2-binary==2.9.5
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
PySocks @ file:///C:/ci/pysocks_1605307512533/work
python-dotenv==0.21.1
pytz==2022.7.1
pywin32==302
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
ruamel-yaml-conda @ file:///C:/ci/ruamel_yaml_1616016898638/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
soupsieve==2.3.2.post1
SQLAlchemy==1.4.37
sqlparse==0.4.3
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
tzdata==2022.7
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
virtualenv==20.19.0
Werkzeug==2.1.2
win-inet-pton @ file:///C:/ci/win_inet_pton_1605306162074/work
wincertstore==0.2
WTForms==3.0.1
zipp==3.8.0
